render inputs in shared code
